+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Added
+- `-T / --tabsize=COLS` option.
+
 ### Changed
 
 - Separators of grid style layout from only space to tab char and space for GNU ls compatibilities.